refactor: ErrorCode 및 Response 리펙토링
1. ErrorCode만 받는 생성자 추가
2. ErrorCode 코드 정리... ( 추후 분리가 필요해보임! )
